requirements.txt
# AI Insights Generator - Requirements
# Updated for 2025 with latest stable versions

# Core Web Framework
Flask==3.1.1
Werkzeug==3.1.3

# CrewAI Multi-Agent System (latest stable)
crewai==0.134.0
crewai-tools==0.48.0

# Language Models & AI APIs
openai==1.77.0
litellm==1.72.0

# Search & Research Tools
tavily-python==0.5.0
google-search-results==2.4.2
serper==0.1.4

# Data Processing & Validation
pydantic==2.11.4
python-dotenv==1.1.0
requests==2.32.3

# Google Cloud Services (optional - for advanced features)
google-cloud-secret-manager==2.21.1
google-cloud-firestore==2.19.0
firebase-admin==6.5.0

# Utilities
uuid==1.30
typing-extensions==4.12.2

# Development & Testing (optional)
pytest==8.4.0
flask-testing==0.8.1

# Production Deployment (optional)
gunicorn==23.0.0
waitress==3.0.2

# Security & Performance
cryptography==44.0.3
cachetools==5.5.2 